0x0C. C - More malloc, free

+++ 0x0C-more_malloc_free/0-main.c
@@ -2,6 +2,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <limits.h>
+#include <stdint.h>
 
 /**
  * main - check the code
@@ -21,7 +22,7 @@ int main(void)
     printf("%p\n", (void *)i);
     f = malloc_checked(sizeof(float) * 100000000);
     printf("%p\n", (void *)f);
-    d = malloc_checked(INT_MAX);
+    d = malloc_checked(SIZE_MAX);
     printf("%p\n", (void *)d);
     free(c);
     free(i);
